requirements.txt
argon2-cffi==20.1.0
crispy-tailwind==0.2.0
Django==2.2.17
django-allauth==0.44.0
django-environ==0.4.5
django-extensions==3.1.0
django-model-utils==4.1.1
gunicorn==20.0.4
Pillow==8.0.1
pre-commit==2.9.2
psycopg2-binary==2.8.6
